backend/requirements.txt
# ================= CORE BACKEND =================
fastapi==0.135.1
uvicorn==0.40.0
sqlalchemy==2.0.46
pymysql==1.1.2
mysql-connector-python==9.5.0

pydantic==2.12.5
pydantic-settings==2.11.0
python-dotenv==1.2.2

flask==3.1.2
flask-cors==6.0.1
flask-sqlalchemy==3.1.1

marshmallow==3.21.1
python-socketio==5.11.2
python-engineio==4.13.1

requests==2.32.5

# ================= DATA / ML (COMPATIBLE WITH PY 3.14) =================
numpy==2.3.3
pandas==2.3.3
scipy==1.16.2
scikit-learn==1.7.2
joblib==1.5.3
threadpoolctl==3.6.0
gdown

# ================= REQUIRED INDIRECT DEPENDENCIES =================
anyio==4.12.1
click==8.3.1
h11==0.16.0
greenlet==3.3.2
typing-extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0
annotated-doc==0.0.4

charset-normalizer==3.4.6
idna==3.11
urllib3==2.6.3
certifi==2026.2.25

jinja2==3.1.6
markupsafe==3.0.3
werkzeug==3.1.6
itsdangerous==2.2.0
blinker==1.9.0

packaging==26.0
python-dateutil==2.9.0.post0
six==1.17.0

bidict==0.23.1
simple-websocket==1.1.0
wsproto==1.3.2
apscheduler==3.10.4
